requirements.txt
# Core dependencies
llama-cpp-python>=0.2.11
pydantic>=2.0.0
pyyaml>=6.0.1
psutil>=5.9.6
click>=8.1.0

# Development tools
pytest>=7.4.0
